toggleHiddenThoughts
